+++ src/authReturn.ts
@@ -1,6 +1,6 @@
 import { parse } from 'simple-query-string';
-import { getAccessTokenFromCode } from './app/oauth/oauth.service';
-import { setToken } from './app/oauth/oauth-token.service';
+import { getAccessTokenFromCode } from './app/bungie-api/oauth';
+import { setToken } from './app/bungie-api/oauth-tokens';
 import { reportException } from './app/utils/exceptions';
 
 function handleAuthReturn() {
